pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.4",
    "django-bootstrap5==24.3",
    "django-filter>=25.1",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.1",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "coverage>=7.10.2",
    "pytest>=8.4.1",
    "pytest-django>=4.11.1",
    "ruff>=0.12.7",
]
